# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nour — Egyptian Fashion & Beauty Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /store/{slug}:
    get:
      operationId: getStorefront
      tags: [storefront]
      summary: Get public merchant storefront by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Merchant storefront data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorefrontResponse"
        "404":
          description: Store not found
...
components:
  schemas:
...
    StorefrontProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        status:
          type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
      required: [id, name, description, price, stock, status]
    StorefrontResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [fashion, cosmetics, both]
        status:
          type: string
        city:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        products:
          type: array
          items:
            $ref: "#/components/schemas/StorefrontProduct"
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        categories:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              nameAr:
                type: string
              parentId:
                type: integer
                nullable: true
              type:
                type: string
                enum: [fashion, cosmetics]
              imageUrl:
                type: string
                nullable: true
              productCount:
                type: integer
            required: [id, name]
        storeConfig:
          type: object
          nullable: true
          additionalProperties: true
      required: [id, name, slug, description, category, status, createdAt, products, totalProducts, totalOrders, categories]
...